Update webclient and split up backend
