+++ .gitignore
@@ -5,3 +5,4 @@ node_modules
 *.code-workspace
 .vscode
 .env
+__pycache__/
