+++ media/components/sdfv/index.html
@@ -389,14 +389,6 @@
             vscode.postMessage({
                 type: 'sdfv.get_current_sdfg',
             });
-            vscode.postMessage({
-                type: 'dace.get_enum',
-                name: 'InstrumentationType',
-            });
-            vscode.postMessage({
-                type: 'dace.get_enum',
-                name: 'AccessType',
-            });
         });
     </script>
 </body>
